Revert "Music: scroll timeline during playback (#56584)"
This reverts commit 096d4a1c7d044a52adacd3b739ade8c7b650f348.
