+++ apps/src/music/views/Timeline.tsx
@@ -1,4 +1,4 @@
-import React, {MouseEvent, useCallback, useRef} from 'react';
+import React, {MouseEvent, useCallback} from 'react';
 import moduleStyles from './timeline.module.scss';
 import classNames from 'classnames';
 import TimelineSampleEvents from './TimelineSampleEvents';
@@ -12,15 +12,12 @@ import {
   setStartPlayheadPosition,
 } from '../redux/musicRedux';
 import {useMusicSelector} from './types';
-import usePlaybackUpdate from './hooks/usePlaybackUpdate';
 
 const barWidth = 60;
 // Leave some vertical space between each event block.
 const eventVerticalSpace = 2;
 // A little room on the left.
 const paddingOffset = 10;
-// Start scrolling the playhead when it's more than this percentage of the way across the timeline area.
-const playheadScrollThreshold = 0.75;
 
 const getEventHeight = (numUniqueRows: number, availableHeight = 110) => {
   // While we might not actually have this many rows to show,
@@ -56,8 +53,6 @@ const Timeline: React.FunctionComponent = () => {
     MIN_NUM_MEASURES,
     useMusicSelector(state => state.music.lastMeasure)
   );
-  const playheadRef = useRef<HTMLDivElement>(null);
-  const timelineRef = useRef<HTMLDivElement>(null);
 
   const positionToUse = isPlaying
     ? currentPlayheadPosition
@@ -110,36 +105,11 @@ const Timeline: React.FunctionComponent = () => {
     dispatch(clearSelectedBlockId());
   }, [dispatch]);
 
-  const scrollPlayheadForward = useCallback(() => {
-    if (!timelineRef.current || !playheadRef.current) {
-      return;
-    }
-
-    const playheadOffset =
-      playheadRef.current.getBoundingClientRect().left -
-      timelineRef.current.getBoundingClientRect().left;
-    const scrollThreshold =
-      timelineRef.current.clientWidth * playheadScrollThreshold;
-    if (playheadOffset > scrollThreshold) {
-      timelineRef.current.scrollBy(playheadOffset - scrollThreshold, 0);
-    }
-  }, [playheadRef]);
-
-  const scrollToPlayhead = useCallback(() => {
-    playheadRef.current?.scrollIntoView();
-  }, [playheadRef]);
-
-  usePlaybackUpdate(scrollPlayheadForward, scrollToPlayhead, scrollToPlayhead);
-
   return (
     <div
       id="timeline"
-      className={classNames(
-        moduleStyles.timeline,
-        isPlaying && moduleStyles.timelinePlaying
-      )}
+      className={moduleStyles.timeline}
       onClick={onTimelineClick}
-      ref={timelineRef}
     >
       <div
         id="timeline-measures-background"
@@ -199,7 +169,6 @@ const Timeline: React.FunctionComponent = () => {
             isPlaying && moduleStyles.playheadPlaying
           )}
           style={{left: paddingOffset + playHeadOffsetInPixels}}
-          ref={playheadRef}
         >
           &nbsp;
         </div>
