+++ apps/src/music/views/timeline.module.scss
@@ -12,10 +12,6 @@
   height: 100%;
   position: relative;
 
-  &Playing {
-    overflow-x: hidden;
-  }
-
   .fullWidthOverlay {
     height: 100%;
     position: absolute;
